requirements.txt
aioitertools==0.7.1
amqp==2.6.1
asgiref==3.3.4
atomate==0.9.9
attrs==20.3.0
automium==0.2.6
automium-web==0.1.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
brewer2mpl==1.4.1
castepxbin==0.1.0
celery==4.4.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.7
custodian==2021.2.8
cycler==0.10.0
decorator==4.4.2
Django==3.2
django-debug-toolbar==3.2.1
django-extensions==3.1.2
django-filter==2.4.0
djangorestframework==3.12.4
dnspython==2.1.0
drf-yasg==1.20.0
emmet==2018.6.7
Faker==8.1.0
fastapi==0.63.0
FireWorks==1.9.7
flake8==3.9.0
flake8-builtins==1.5.3
flake8-docstrings==1.6.0
flake8-import-order==0.18.1
flake8-quotes==3.2.0
Flask==1.1.2
flask-paginate==0.8.1
flower==0.9.7
future==0.18.2
gunicorn==20.1.0
h5py==3.2.1
humanize==3.4.1
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
kiwisolver==1.3.1
kombu==4.6.11
latexcodec==2.0.1
legacy==0.1.6
maggma==0.26.0
MarkupSafe==1.1.1
matminer==0.6.5
matplotlib==3.4.1
mccabe==0.6.1
mongogrant==0.3.1
mongomock==3.22.1
monty==2021.3.3
mpmath==1.2.1
msgpack==1.0.2
networkx==2.5.1
numpy==1.20.2
packaging==20.9
palettable==3.3.0
pandas==1.2.4
paramiko==2.7.2
phonopy==2.9.3
Pillow==8.2.0
Pint==0.17
plotly==4.14.3
pluggy==0.13.1
postgres==3.0.0
prettyplotlib==0.1.7
prometheus-client==0.8.0
prompt-toolkit==3.0.18
psycopg2-binary==2.9.1
psycopg2-pool==1.1
py==1.10.0
pybtex==0.24.0
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.1
pydash==5.0.0
pydocstyle==6.0.0
pyflakes==2.3.1
pymatgen==2022.0.5
pymatgen-diffusion==2020.10.8
pymongo==3.11.3
PyNaCl==1.4.0
pynng==0.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.4.0
pytest-mock==3.6.1
python-dateutil==2.8.1
python-memcached==1.59
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
retrying==1.3.3
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.1
scipy==1.6.2
seekpath==2.0.1
sentinels==1.0.0
sentry-sdk==1.0.0
six==1.15.0
sniffio==1.2.0
snowballstemmer==2.1.0
soupsieve==2.2.1
spglib==1.16.1
sqlparse==0.4.1
sshtunnel==0.4.0
starlette==0.13.6
sumo==2.2.3
swagger-spec-validator==2.7.3
sympy==1.8
tabulate==0.8.9
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.60.0
typing-extensions==3.7.4.3
uncertainties==3.1.5
uritemplate==3.0.1
urllib3==1.26.4
vine==1.3.0
wcwidth==0.2.5
Werkzeug==1.0.1
